construct alphabet
The alphabet is constructed by walking down a tree of letters. The
tree is constructed in the following manner.
Imaging a node 'A' if we add a letter 'B' we add it to the left if and
only if 'BA' occurs more often then 'AB'. If 'AB' occurs more often we
add it to the right, otherwise we do not care.